Add base git classes
