+++ txgh/lib/txgh.rb
@@ -7,10 +7,13 @@ module Txgh
   autoload :DiffCalculator,        'txgh/diff_calculator'
   autoload :EmptyResourceContents, 'txgh/empty_resource_contents'
   autoload :Events,                'txgh/events'
+  autoload :GitApi,                'txgh/git_api'
   autoload :GithubApi,             'txgh/github_api'
   autoload :GitlabApi,             'txgh/gitlab_api'
+  autoload :GitRepo,               'txgh/git_repo'
   autoload :GithubRepo,            'txgh/github_repo'
   autoload :GitlabRepo,            'txgh/gitlab_repo'
+  autoload :GitStatus,             'txgh/git_status'
   autoload :GithubStatus,          'txgh/github_status'
   autoload :GitlabStatus,          'txgh/gitlab_status'
   autoload :MergeCalculator,       'txgh/merge_calculator'
