+++ txgh/lib/txgh/config/providers.rb
@@ -2,6 +2,7 @@ module Txgh
   module Config
     module Providers
       autoload :FileProvider,   'txgh/config/providers/file_provider'
+      autoload :GitProvider, 'txgh/config/providers/git_provider'
       autoload :GithubProvider, 'txgh/config/providers/github_provider'
       autoload :GitlabProvider, 'txgh/config/providers/gitlab_provider'
       autoload :RawProvider,    'txgh/config/providers/raw_provider'
