+++ txgh/lib/txgh/config/providers/gitlab_provider.rb
@@ -1,13 +1,7 @@
 module Txgh
   module Config
     module Providers
-      class GitlabProvider < Txgh::Config::Providers::GithubProvider
-        def initialize(payload, parser, options = {})
-          @payload = payload
-          @parser = parser
-          @ref = options[:ref]
-          @gitlab_repo = options[:gitlab_repo]
-        end
+      class GitlabProvider < GitProvider
 
         private
 
@@ -15,15 +9,6 @@ module Txgh
           gitlab_repo.api.download(payload, ref)[:content]
         rescue ::Gitlab::Error::NotFound
           raise Txgh::GitConfigNotFoundError, "Config file #{payload} not found in #{ref}"
-        end
-
-        def gitlab_repo
-          unless @gitlab_repo
-            raise TxghError,
-              "TX_CONFIG specified a file from git but did not provide a repo."
-          end
-
-          @gitlab_repo
         end
       end
     end
