+++ txgh/lib/txgh/git_api.rb
@@ -1,13 +1,18 @@
-require 'octokit'
+require 'base64'
 
 module Txgh
-  class GithubApi < GitApi
+  class GitApi
     class << self
-      def create_from_credentials(login, access_token, repo_name)
-        create_from_client(
-          Octokit::Client.new(login: login, access_token: access_token), repo_name
-        )
+      def create_from_client(client, repo_name)
+        new(client, repo_name)
       end
+    end
+
+    attr_reader :client, :repo_name
+
+    def initialize(client, repo_name)
+      @client = client
+      @repo_name = repo_name
     end
 
     def tree(sha)
