+++ txgh/spec/config/config_pair_spec.rb
@@ -1,20 +1,18 @@
 require 'spec_helper'
 require 'helpers/standard_txgh_setup'
 
-include Txgh::Config
-
-describe ConfigPair do
+describe Txgh::Config::ConfigPair do
   include StandardTxghSetup
 
   let(:config) do
-    ConfigPair.new(project_config, github_config)
+    described_class.new(project_config, github_config)
   end
 
   describe '#git_repo' do
     context 'Github' do
       it 'instantiates a github repo with the right config' do
         repo = config.git_repo
-        expect(repo).to be_a(GithubRepo)
+        expect(repo).to be_a(Txgh::GithubRepo)
         expect(repo.name).to eq(github_repo_name)
         expect(repo.branch).to eq(branch)
       end
@@ -22,12 +20,12 @@ describe ConfigPair do
 
     context 'Gitlab' do
       let(:config) do
-        ConfigPair.new(project_config, gitlab_config)
+        described_class.new(project_config, gitlab_config)
       end
 
       it 'instantiates a github repo with the right config' do
         repo = config.git_repo
-        expect(repo).to be_a(GitlabRepo)
+        expect(repo).to be_a(Txgh::GitlabRepo)
         expect(repo.name).to eq(gitlab_repo_name)
         expect(repo.branch).to eq(branch)
       end
@@ -37,7 +35,7 @@ describe ConfigPair do
   describe '#transifex_project' do
     it 'instantiates a transifex project with the right config' do
       project = config.transifex_project
-      expect(project).to be_a(TransifexProject)
+      expect(project).to be_a(Txgh::TransifexProject)
       expect(project.name).to eq(project_name)
       expect(tx_config.resources.first.resource_slug).to eq(resource_slug)
     end
@@ -46,7 +44,7 @@ describe ConfigPair do
   describe '#transifex_api' do
     it 'instantiates an API instance' do
       api = config.transifex_api
-      expect(api).to be_a(TransifexApi)
+      expect(api).to be_a(Txgh::TransifexApi)
       expect(api.connection.headers).to include('Authorization')
     end
   end
@@ -54,7 +52,7 @@ describe ConfigPair do
   describe '#github_api' do
     it 'instantiates a github API instance' do
       api = config.github_api
-      expect(api).to be_a(GithubApi)
+      expect(api).to be_a(Txgh::GithubApi)
       expect(api.client.login).to eq(github_config['api_username'])
       expect(api.client.access_token).to eq(github_config['api_token'])
     end
@@ -62,12 +60,12 @@ describe ConfigPair do
 
   describe '#gitlab_api' do
     let(:config) do
-      ConfigPair.new(project_config, gitlab_config)
+      described_class.new(project_config, gitlab_config)
     end
 
     it 'instantiates a gitlab API instance' do
       api = config.gitlab_api
-      expect(api).to be_a(GitlabApi)
+      expect(api).to be_a(Txgh::GitlabApi)
       expect(api.client.private_token).to eq(gitlab_config['api_token'])
     end
   end
